refactor(connectors): split management call queries
Move the six management-call queries into their own per-resource module while preserving package-root re-export identity for all existing callers. Add structural coverage for the module placement and exports.\n\nCo-Authored-By: OpenAI Codex <codex@openai.com>
